requirements.txt
PyQt6==6.7.0
PyQt6-WebEngine==6.7.0
openai==1.30.0
Pillow==10.3.0
pyautogui==0.9.54
mss==9.0.1
